+++ backend/src/main/java/es/iesdeteis/gestorcitas/repository/HorarioTallerRepository.java
@@ -3,5 +3,5 @@ package es.iesdeteis.gestorcitas.repository;
 import es.iesdeteis.gestorcitas.model.HorarioTaller;
 import org.springframework.data.repository.CrudRepository;
 
-public interface HorarioTallerRepository extends CrudRepository<HorarioTaller, Integer> {
+public interface HorarioTallerRepository extends CrudRepository<HorarioTaller, Long> {
 }
