+++ backend/src/main/java/es/iesdeteis/gestorcitas/repository/TallerRepository.java
@@ -3,5 +3,5 @@ package es.iesdeteis.gestorcitas.repository;
 import es.iesdeteis.gestorcitas.model.Taller;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TallerRepository extends CrudRepository<Taller, Integer> {
+public interface TallerRepository extends CrudRepository<Taller, Long> {
 }
